move: Update import path for useMainContext
